@@ -1,17 +1,20 @@
 """Constants for the OJMicroline component."""
+
 DOMAIN = "ojmicroline_thermostat"
 MANUFACTURER = "OJ Electronics"
 INTEGRATION_NAME = "OJ Microline Thermostat"
-CONFIG_FLOW_VERSION = 1
+CONFIG_FLOW_VERSION = 2
 
 API_TIMEOUT = 30
 UPDATE_INTERVAL = 60
 
+CONF_MODEL = "model"
 CONF_CUSTOMER_ID = "customer_id"
-CONF_DEFAULT_HOST = "ocd5.azurewebsites.net"
-CONF_DEFAULT_CUSTOMER_ID = 99
 CONF_USE_COMFORT_MODE = "use_comfort_mode"
 CONF_COMFORT_MODE_DURATION = "comfort_mode_duration"
+
+MODEL_WD5_SERIES = "WD5 series"
+MODEL_WG4_SERIES = "WG4 series"
 
 PRESET_VACATION = "Vacation"
 PRESET_FROST_PROTECTION = "Frost Protection"
