@@ -1,4 +1,5 @@
 """Support for OJ Microline entities."""
+
 from __future__ import annotations
 
 from typing import Any
@@ -16,23 +17,25 @@ class OJMicrolineEntity(CoordinatorEntity[OJMicrolineDataUpdateCoordinator], Ent
     idx: str
 
     def __init__(self, coordinator: OJMicrolineDataUpdateCoordinator, idx: str) -> None:
-        """
-        Initialise the entity.
+        """Initialise the entity.
 
         Args:
+        ----
             coordinator: The data coordinator updating the models.
             idx: The identifier for this entity.
+
         """
         super().__init__(coordinator)
         self.idx = idx
 
     @property
     def device_info(self) -> dict[str, Any]:
-        """
-        Return information to link this entity with the correct device.
+        """Return information to link this entity with the correct device.
 
-        Returns:
+        Returns
+        -------
             The device identifiers to make sure the entity is attached
             to the correct device.
+
         """
         return {"identifiers": {(DOMAIN, self.idx)}}
